requirements.txt

PyYAML==6.0.2
Jinja2==3.1.4
jinja2-strcase==0.0.2
argparse==1.4.0
stringcase==1.2.0
protobuf==6.33.0